+++ crates/miden-lib/src/transaction/inputs.rs
@@ -104,6 +104,7 @@ impl TransactionAdviceInputs {
     ///     number_of_input_notes,
     ///     TX_SCRIPT_ROOT,
     ///     TX_SCRIPT_ARGS_KEY,
+    ///     AUTH_ARGS,
     /// ]
     fn build_stack(
         &mut self,
@@ -147,7 +148,10 @@ impl TransactionAdviceInputs {
         // --- number of notes, script root and args key ----------------------
         self.extend_stack([Felt::from(tx_inputs.input_notes().num_notes())]);
         self.extend_stack(tx_args.tx_script().map_or(Word::empty(), |script| script.root()));
-        self.extend_stack(tx_args.tx_script_arg());
+        self.extend_stack(tx_args.tx_script_args());
+
+        // --- auth procedure args -------------------------------------------
+        self.extend_stack(tx_args.auth_args());
     }
 
     // BLOCKCHAIN INJECTIONS