+++ crates/miden-objects/src/transaction/tx_args.rs
@@ -32,13 +32,18 @@ use crate::{
 /// - Advice inputs: provides data needed by the runtime, like the details of public output notes.
 /// - Foreign account inputs: provides foreign account data that will be used during the foreign
 ///   procedure invocation (FPI).
+/// - Auth arguments: data put onto the stack right before authentication procedure execution. If
+///   this argument is not specified, the [`EMPTY_WORD`] would be used as a default value. If the
+///   [AdviceInputs] are propagated with some user defined map entries, this argument could be used
+///   as a key to access the corresponding value.
 #[derive(Clone, Debug, Default, PartialEq, Eq)]
 pub struct TransactionArgs {
     tx_script: Option<TransactionScript>,
-    tx_script_arg: Word,
+    tx_script_args: Word,
     note_args: BTreeMap<NoteId, Word>,
     advice_inputs: AdviceInputs,
     foreign_account_inputs: Vec<AccountInputs>,
+    auth_args: Word,
 }
 
 impl TransactionArgs {
@@ -50,10 +55,11 @@ impl TransactionArgs {
     pub fn new(advice_map: AdviceMap, foreign_account_inputs: Vec<AccountInputs>) -> Self {
         Self {
             tx_script: None,
-            tx_script_arg: EMPTY_WORD,
+            tx_script_args: EMPTY_WORD,
             note_args: Default::default(),
             advice_inputs: AdviceInputs::default().with_map(advice_map),
             foreign_account_inputs,
+            auth_args: EMPTY_WORD,
         }
     }
 
@@ -76,10 +82,10 @@ impl TransactionArgs {
     pub fn with_tx_script_and_arg(
         mut self,
         tx_script: TransactionScript,
-        tx_script_arg: Word,
+        tx_script_args: Word,
     ) -> Self {
         self.tx_script = Some(tx_script);
-        self.tx_script_arg = tx_script_arg;
+        self.tx_script_args = tx_script_args;
         self
     }
 
@@ -90,6 +96,13 @@ impl TransactionArgs {
     #[must_use]
     pub fn with_note_args(mut self, note_args: BTreeMap<NoteId, Word>) -> Self {
         self.note_args = note_args;
+        self
+    }
+
+    /// Returns new [TransactionArgs] instantiated with the provided auth arguments.
+    #[must_use]
+    pub fn with_auth_args(mut self, auth_args: Word) -> Self {
+        self.auth_args = auth_args;
         self
     }
 
@@ -108,8 +121,8 @@ impl TransactionArgs {
     /// transaction script execution. Notice that the corresponding map entry should be provided
     /// separately during the creation with the [`TransactionArgs::new`] or using the
     /// [`TransactionArgs::extend_advice_map`] method.
-    pub fn tx_script_arg(&self) -> Word {
-        self.tx_script_arg
+    pub fn tx_script_args(&self) -> Word {
+        self.tx_script_args
     }
 
     /// Returns a reference to a specific note argument.
@@ -133,6 +146,17 @@ impl TransactionArgs {
             .iter()
             .map(|acc| acc.code().commitment())
             .collect()
+    }
+
+    /// Returns a reference to the authentication procedure argument, or [`EMPTY_WORD`] if the
+    /// argument was not specified.
+    ///
+    /// This argument could be potentially used as a key to access the advice map during the
+    /// transaction script execution. Notice that the corresponding map entry should be provided
+    /// separately during the creation with the [`TransactionArgs::new`] or using the
+    /// [`TransactionArgs::extend_advice_map`] method.
+    pub fn auth_args(&self) -> Word {
+        self.auth_args
     }
 
     // STATE MUTATORS
@@ -197,27 +221,30 @@ impl TransactionArgs {
 impl Serializable for TransactionArgs {
     fn write_into<W: ByteWriter>(&self, target: &mut W) {
         self.tx_script.write_into(target);
-        self.tx_script_arg.write_into(target);
+        self.tx_script_args.write_into(target);
         self.note_args.write_into(target);
         self.advice_inputs.write_into(target);
         self.foreign_account_inputs.write_into(target);
+        self.auth_args.write_into(target);
     }
 }
 
 impl Deserializable for TransactionArgs {
     fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
         let tx_script = Option::<TransactionScript>::read_from(source)?;
-        let tx_script_arg = Word::read_from(source)?;
+        let tx_script_args = Word::read_from(source)?;
         let note_args = BTreeMap::<NoteId, Word>::read_from(source)?;
         let advice_inputs = AdviceInputs::read_from(source)?;
         let foreign_account_inputs = Vec::<AccountInputs>::read_from(source)?;
+        let auth_args = Word::read_from(source)?;
 
         Ok(Self {
             tx_script,
-            tx_script_arg,
+            tx_script_args,
             note_args,
             advice_inputs,
             foreign_account_inputs,
+            auth_args,
         })
     }
 }