+++ crates/miden-testing/src/kernel_tests/block/proposed_block_success.rs
@@ -14,7 +14,10 @@ use super::utils::{
     generate_fungible_asset, generate_tracked_note_with_asset, generate_tx_with_expiration,
     generate_tx_with_unauthenticated_notes, generate_untracked_note, setup_chain,
 };
-use crate::ProvenTransactionExt;
+use crate::{
+    ProvenTransactionExt,
+    kernel_tests::block::utils::{generate_account_with_conditional_auth, generate_conditional_tx},
+};
 
 /// Tests that we can build empty blocks.
 #[test]
@@ -243,6 +246,45 @@ fn proposed_block_with_batch_at_expiration_limit() -> anyhow::Result<()> {
     // at block 3 (due to tx1) should still be accepted into the block.
     let block_inputs = chain.get_block_inputs(&batches)?;
     ProposedBlock::new(block_inputs.clone(), batches.clone())?;
+
+    Ok(())
+}
+
+/// Tests that a NOOP transaction with state commitments X -> X against account A can appear
+/// in one batch while another batch contains a state-updating transaction with state commitments X
+/// -> Y against the same account A. Both batches are in the same block.
+#[test]
+fn noop_tx_and_state_updating_tx_against_same_account_in_same_block() -> anyhow::Result<()> {
+    let TestSetup { mut chain, .. } = setup_chain(0);
+
+    let account0 = generate_account_with_conditional_auth(&mut chain);
+
+    let noop_tx = generate_conditional_tx(&mut chain, account0.id(), false);
+    let state_updating_tx = generate_conditional_tx(&mut chain, noop_tx.clone(), true);
+
+    // sanity check: NOOP transaction's init and final commitment should be the same.
+    assert_eq!(noop_tx.initial_account().commitment(), noop_tx.final_account().commitment());
+    // sanity check: State-updating transaction's init and final commitment should *not* be the
+    // same.
+    assert_ne!(
+        state_updating_tx.initial_account().commitment(),
+        state_updating_tx.final_account().commitment()
+    );
+
+    let tx0 = ProvenTransaction::from_executed_transaction_mocked(noop_tx);
+    let tx1 = ProvenTransaction::from_executed_transaction_mocked(state_updating_tx);
+
+    let batch0 = generate_batch(&mut chain, vec![tx0]);
+    let batch1 = generate_batch(&mut chain, vec![tx1.clone()]);
+
+    let batches = vec![batch0.clone(), batch1.clone()];
+
+    let block_inputs = chain.get_block_inputs(&batches)?;
+    let block = ProposedBlock::new(block_inputs, batches.clone())?;
+
+    let (_, update) = block.updated_accounts().iter().next().unwrap();
+    assert_eq!(update.initial_state_commitment(), account0.commitment());
+    assert_eq!(update.final_state_commitment(), tx1.account_update().final_state_commitment());
 
     Ok(())
 }