+++ crates/miden-testing/src/kernel_tests/block/utils.rs
@@ -2,14 +2,16 @@ use std::{collections::BTreeMap, vec, vec::Vec};
 
 use miden_lib::{note::create_p2id_note, transaction::TransactionKernel};
 use miden_objects::{
-    Felt, Word,
+    Felt, ONE, Word, ZERO,
     account::{Account, AccountId, AccountStorageMode},
     asset::{Asset, FungibleAsset},
     batch::ProvenBatch,
     block::BlockNumber,
     crypto::rand::RpoRandomCoin,
     note::{Note, NoteId, NoteTag, NoteType},
-    testing::{account_component::AccountMockComponent, note::NoteBuilder},
+    testing::{
+        account_component::AccountMockComponent, account_id::ACCOUNT_ID_SENDER, note::NoteBuilder,
+    },
     transaction::{ExecutedTransaction, OutputNote, ProvenTransaction, TransactionScript},
 };
 use rand::{Rng, SeedableRng, rngs::SmallRng};
@@ -30,6 +32,17 @@ pub fn generate_account(chain: &mut MockChain) -> Account {
         );
     chain
         .add_pending_account_from_builder(Auth::IncrNonce, account_builder, AccountState::Exists)
+        .expect("failed to add pending account from builder")
+}
+
+pub fn generate_account_with_conditional_auth(chain: &mut MockChain) -> Account {
+    let account_builder = Account::builder(rand::rng().random())
+        .storage_mode(AccountStorageMode::Public)
+        .with_component(
+            AccountMockComponent::new_with_empty_slots(TransactionKernel::assembler()).unwrap(),
+        );
+    chain
+        .add_pending_account_from_builder(Auth::Conditional, account_builder, AccountState::Exists)
         .expect("failed to add pending account from builder")
 }
 
@@ -108,6 +121,39 @@ pub fn generate_tx_with_authenticated_notes(
 ) -> ProvenTransaction {
     let executed_tx = generate_executed_tx_with_authenticated_notes(chain, account_id, notes);
     ProvenTransaction::from_executed_transaction_mocked(executed_tx)
+}
+
+/// Generates a transaction, which depending on the `modify_storage` flag, does the following:
+/// - if `modify_storage` is true, it increments the storage item of the account.
+/// - if `modify_storage` is false, it does nothing (NOOP).
+///
+/// To make this transaction (always) non-empty, it consumes one "noop note", which does nothing.
+pub fn generate_conditional_tx(
+    chain: &mut MockChain,
+    input: impl Into<TxContextInput>,
+    modify_storage: bool,
+) -> ExecutedTransaction {
+    let noop_note = NoteBuilder::new(ACCOUNT_ID_SENDER.try_into().unwrap(), &mut rand::rng())
+        .build(&TransactionKernel::assembler())
+        .expect("failed to create the noop note");
+    chain.add_pending_note(OutputNote::Full(noop_note.clone()));
+    chain.prove_next_block().unwrap();
+
+    let auth_args = [
+        if modify_storage { ONE } else { ZERO }, // increment nonce if modify_storage is true
+        Felt::new(99),
+        Felt::new(98),
+        Felt::new(97),
+    ];
+
+    let tx_context = chain
+        .build_tx_context(input.into(), &[noop_note.id()], &[])
+        .unwrap()
+        .extend_input_notes(vec![noop_note])
+        .auth_args(auth_args.into())
+        .build()
+        .unwrap();
+    tx_context.execute().unwrap()
 }
 
 /// Generates a transaction that expires at the given block number.