+++ crates/miden-testing/src/kernel_tests/tx/mod.rs
@@ -18,6 +18,7 @@ mod test_account_delta;
 mod test_account_interface;
 mod test_asset;
 mod test_asset_vault;
+mod test_auth;
 mod test_epilogue;
 mod test_faucet;
 mod test_fpi;