+++ crates/miden-testing/src/tx_context/builder.rs
@@ -68,9 +68,10 @@ pub struct TransactionContextBuilder {
     foreign_account_inputs: Vec<AccountInputs>,
     input_notes: Vec<Note>,
     tx_script: Option<TransactionScript>,
-    tx_script_arg: Word,
+    tx_script_args: Word,
     note_args: BTreeMap<NoteId, Word>,
     transaction_inputs: Option<TransactionInputs>,
+    auth_args: Word,
 }
 
 impl TransactionContextBuilder {
@@ -82,12 +83,13 @@ impl TransactionContextBuilder {
             input_notes: Vec::new(),
             expected_output_notes: Vec::new(),
             tx_script: None,
-            tx_script_arg: EMPTY_WORD,
+            tx_script_args: EMPTY_WORD,
             authenticator: None,
             advice_inputs: Default::default(),
             transaction_inputs: None,
             note_args: BTreeMap::new(),
             foreign_account_inputs: vec![],
+            auth_args: EMPTY_WORD,
         }
     }
 
@@ -124,10 +126,11 @@ impl TransactionContextBuilder {
             expected_output_notes: Vec::new(),
             advice_inputs: Default::default(),
             tx_script: None,
-            tx_script_arg: EMPTY_WORD,
+            tx_script_args: EMPTY_WORD,
             transaction_inputs: None,
             note_args: BTreeMap::new(),
             foreign_account_inputs: vec![],
+            auth_args: EMPTY_WORD,
         }
     }
 
@@ -223,8 +226,14 @@ impl TransactionContextBuilder {
     }
 
     /// Set the transaction script argument
-    pub fn tx_script_arg(mut self, tx_script_arg: Word) -> Self {
-        self.tx_script_arg = tx_script_arg;
+    pub fn tx_script_args(mut self, tx_script_args: Word) -> Self {
+        self.tx_script_args = tx_script_args;
+        self
+    }
+
+    /// Set the desired auth arguments
+    pub fn auth_args(mut self, auth_args: Word) -> Self {
+        self.auth_args = auth_args;
         self
     }
 
@@ -291,10 +300,12 @@ impl TransactionContextBuilder {
             .with_note_args(self.note_args);
 
         let mut tx_args = if let Some(tx_script) = self.tx_script {
-            tx_args.with_tx_script_and_arg(tx_script, self.tx_script_arg)
+            tx_args.with_tx_script_and_arg(tx_script, self.tx_script_args)
         } else {
             tx_args
         };
+
+        tx_args = tx_args.with_auth_args(self.auth_args);
 
         tx_args.extend_advice_inputs(self.advice_inputs.clone());
         tx_args.extend_output_note_recipients(self.expected_output_notes.clone());