+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - [BREAKING] Renamed `{TransactionId, NoteId, Nullifier}::inner` -> `as_word` ([#1571](https://github.com/0xMiden/miden-base/pull/1571)).
 - Replace `MerklePath` with `SparseMerklePath` in `NoteInclusionProof` ([#1572](https://github.com/0xMiden/miden-base/pull/1572)) .
 - [BREAKING] Split `TransactionHost` into `TransactionProverHost` and `TransactionExecutorHost` ([#1581](https://github.com/0xMiden/miden-base/pull/1581)).
+- Add arguments to the auth procedure ([#1501](https://github.com/0xMiden/miden-base/pull/1501)).
 
 ## 0.10.0 (2025-07-08)